PROBLEM: VBox causes USB hard drive to dismount
When a virtual machine is configured to see a USB hard drive,
at least under Mac OS X, any mounted file systems on the
external hard drive will go away. This means that the
virtual machine files cannot be stored on the USB drive that
the virtual machine is trying to see.